refactor(lowlevel): split signature calculation to separate file
