+++ lowlevel/encryptor.go
@@ -3,30 +3,10 @@ package lowlevel
 import (
 	"crypto/aes"
 	"crypto/cipher"
-	"crypto/sha1"
 	"encoding/base64"
 	"encoding/binary"
 	"errors"
-	"fmt"
-	"sort"
 )
-
-func makeDevMsgSignature(paramValues ...string) string {
-	tmp := make([]string, len(paramValues))
-	for i, x := range paramValues {
-		tmp[i] = x
-	}
-
-	sort.Strings(tmp)
-
-	state := sha1.New()
-	for _, x := range tmp {
-		_, _ = state.Write([]byte(x))
-	}
-
-	result := state.Sum(nil)
-	return fmt.Sprintf("%x", result)
-}
 
 type WorkwxPayload struct {
 	Msg       []byte
